server-side/backend/requirements.txt
django==5.2.8
djangorestframework==3.16.1
psycopg2-binary==2.9.11
dotenv==0.9.9
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
Pillow==10.4.0
drf-spectacular==0.27.2
django-jazzmin==3.0.1
channels==4.0.0
channels-redis==4.2.0
redis==5.0.1
daphne==4.1.0